third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/object-src-none-allowed.html

<!DOCTYPE html>
<html>
<head>
<script>
if (window.testRunner) {
  testRunner.dumpAsText();
  testRunner.dumpChildFrames();
}
</script>
</head>
<body>
  <iframe src="http://127.0.0.1:8000/security/contentSecurityPolicy/resources/echo-object-data.pl?plugin=data:application/x-blink-test-plugin,&log=PASS!&csp=img-src%20'none'"></iframe>
  <script src="/js-test-resources/plugin.js"></script>
  <script>startAfterLoadAndFinish(null, document.querySelector('iframe'));</script>
</body>
</html>
