third_party/blink/web_tests/platform/generic/http/tests/security/contentSecurityPolicy/object-src-none-allowed-expected.txt

CONSOLE MESSAGE: Blink Test Plugin: initializing


--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS!
